requirements.txt
# requirements.txt

blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
PyMuPDF==1.25.1
Werkzeug==3.1.3
